Use a class for parallel computing
The problems is that the `Descriptor` of each sub-process has to be the same.
Hence, a class is created to hold the string form of the descriptor and the
actual `Descriptor` object is created on-demand inside the scope of subprocess.
This way, we avoid hardcoding the descriptor definition in the python file.
A click commandline interface is also added.